requirements.txt
google-api-python-client>=2.165
google-auth>=2.48
google-auth-oauthlib>=1.2
requests>=2.28
python-dotenv>=1.0
